# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.CreateSchedulePayload:
    properties:
      end_date:
        description: YYYY-MM-DD
        type: string
      start_date:
        description: YYYY-MM-DD
        type: string
    required:
    - end_date
    - start_date
    type: object
...
  main.UpdateSchedulePayload:
    properties:
      end_date:
        description: YYYY-MM-DD
        type: string
      start_date:
        description: YYYY-MM-DD
        type: string
    type: object
...
  store.Schedule:
    properties:
      created_at:
        type: string
      end_date:
        description: DateOnly auto-normalizes to YYYY-MM-DD
        type: string
      id:
        type: integer
      published_at:
        type: string
      restaurant_id:
        type: integer
      start_date:
        description: DateOnly auto-normalizes to YYYY-MM-DD
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Restaurant Employee Scheduling Application (RESA)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RESA API
paths:
...
  /restaurants/{restaurant_id}/schedules:
    get:
      consumes:
      - application/json
      description: Fetches all schedules for a restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Schedule'
            type: array
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Lists restaurant's schedules
      tags:
      - schedule
    post:
      consumes:
      - application/json
      description: Creates a schedule for a restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Schedule payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateSchedulePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Schedule'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Creates a schedule
      tags:
      - schedule
  /restaurants/{restaurant_id}/schedules/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a schedule by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Deletes a schedule
      tags:
      - schedule
    get:
      consumes:
      - application/json
      description: Fetches a schedule by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Schedule'
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetches a schedule
      tags:
      - schedule
    patch:
      consumes:
      - application/json
      description: Updates a schedule by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      - description: Schedule payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateSchedulePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Schedule'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Updates a schedule
      tags:
      - schedule
  /restaurants/{restaurant_id}/schedules/{id}/publish:
    post:
      consumes:
      - application/json
      description: Publishes a schedule to make it available to employees
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Publishes a schedule
      tags:
      - schedule
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"